Annotate DNA loops in a Hi-C plot

Annotate DNA loops in a Hi-C plot

bb_annoLoops(hicPlot, bedpeData)

Arguments

hicPlot
Hi-C plot object from bb_plotHicSquare or bb_plotHicTriangle on which to annotate loops.

bedpeData
A string specifying the BEDPE file path or a dataframe in BEDPE format specifying loop positions.

type
Character value specifying type of annotation. Default value is type = "box". Options are:

"box": Boxes are drawn around each loop.

"circle": Circles are drawn around each loop.

"arrow": Arrows are drawn pointing to each loop.

half
Character value specifying which half of hic plots to annotate. Triangle Hi-C plots will always default to the entirety of the triangular plot. Default value is half = "inherit". Options are:

"inherit": Loops will be annotated on the half inherited by the input Hi-C plot.

"both": Loops will be annotated on both halves of the diagonal of a square Hi-C plot.

"top": Loops will be annotated on the upper diagonal half of a square Hi-C plot.

"bottom": Loops will be annotated ont the bottom diagonal half of a square Hi-C plot.

shift
Numeric specifying the number of pixels on either end of loop in a box or circle. Numeric specifying number of pixels for the length of an arrow.

params
An optional bb_assembly object containing relevant function parameters.

...
Additional grid graphical parameters. See gpar.

Value

Returns a bb_loops object containing relevant genomic region, placement, and grob information.

Examples

## Load Hi-C data and BEDPE data
data("bb_hicData")
data("bb_bedpeData")

## Create BentoBox page
bb_pageCreate(width = 3, height = 3, default.units = "inches", xgrid = 0, ygrid = 0)

## Plot and place a square Hi-C plot
hicPlot <- bb_plotHicSquare(hicData = bb_hicData, resolution = 10000, zrange = c(0, 70),
                            chrom = "chr21", chromstart = 28000000, chromend = 30300000,
                            x = 0, y = 0, width = 3, height = 3, just = c("left", "top"), default.units = "inches")

## Annotate loops of Hi-C plot
bb_annoLoops(hicPlot = hicPlot, bedpeData = bb_bedpeData, type = "box", half = "both")