Add an x-axis to a plot

Add an x-axis to a plot

bb_annoXaxis(plot)

Arguments

plot
Plot object to annotate with x-axis.

at
A numeric vector of x-value locations for tick marks.

label
A logical value indicating whether to draw the labels on the tick marks, or an expression or character vector which specify the labels to use. If not logical, must be the same length as the at argument.

main
A logical value indicating whether to draw the x-axis at the bottom of the plot. Default value is main = TRUE. Options are:

TRUE: x-axis is drawn at the bottom of the plot.

FALSE: x-axis is drawn at the top of the plot.

gp
Grid graphical parameters. See gpar.

params
An optional bb_assembly object containing relevant function parameters.

Value

Returns a bb_xaxis object containing relevant grob information.

Examples

## Load transcript information
library("TxDb.Hsapiens.UCSC.hg19.knownGene")
library("org.Hs.eg.db")

## Create BentoBox page
bb_pageCreate(width = 3, height = 2, default.units = "inches", xgrid = 0, ygrid = 0)

## Plot gene transcripts
transcriptPlot <- bb_plotTranscripts(chrom = "chr1", chromstart = 1000000, chromend = 2000000,
                                     x = 0, y = 0, width = 3, height = 1.5, just = c("left", "top"), default.units = "inches")

## Add standard x-axis to transcript plot
bb_annoXaxis(plot = transcriptPlot, at = c(1000000, 1250000, 1500000, 1750000, 2000000), gp = gpar(fontsize = 8))