Make a bb_assembly object for alternate TxDb, OrgDb,and BSgenome genomic annotation packages

Make a bb_assembly object for alternate TxDb, OrgDb,and BSgenome genomic annotation packages

bb_assembly(Genome, TxDb, OrgDb)

Arguments

Genome
String indicating the name of the genome assembly.

TxDb
String of the desired TxDb package name.

OrgDb
String of the desired OrgDb package name.

gene.id.column
String of the TxDb column name that refers to the given TxDb gene IDs. Default value is gene.id.column = "ENTREZID".

display.column
String of the OrgDb column name that refers to the type of gene symbol to be displayed in plots. Default value is display.column = "SYMBOL".

BSgenome
String of the desired BSgenome package name.

Value

Returns a bb_assembly object containing all input parameters.

See also

Examples

## Create a custom bb_assembly object for hg38/GRCh38 packages
newAssembly <- bb_assembly(Genome = "hg38_GRCh38",
                           TxDb = "TxDb.Hsapiens.UCSC.hg38.knownGene",
                           OrgDb = "org.Hs.eg.db",
                           BSgenome = "BSgenome.Hsapiens.NCBI.GRCh38")