Plot genomic coordinates along the x or y-axis of a BentoBox plot

Plot genomic coordinates along the x or y-axis of a BentoBox plot

bb_plotGenomeLabel(plot, x, y,
                   just = c("left", "top"),
                   default.units = "inches")
bb_plotGenomeLabel(chrom, chromstart, chromend,
                   x, y, length,
                   just = c("left", "top"),
                   default.units = "inches")

Arguments

plot
BentoBox plot object to add genome label.

chrom
If not specifying plot, chromosome of genome label, as a string.

chromstart
If not specifying plot, integer start of genome label.

chromend
If not specifying plot, integer end of genome label.

assembly
Default genome assembly as a string or a bb_assembly object. assembly will be inherited from plot, if provided.

fontsize
A numeric specifying text fontsize in points. Default value is fontsize = 10.

fontcolor
A character value indicating the color for text. Default value is fontcolor = "black".

linecolor
A character value indicating the color of the genome label axis. Default value is linecolor = "black".

scale
A character value indicating the scale of the coordinates along the genome label. Default value is scale = "bp". Options are:

"bp": base pairs.

"Kb": kilobase pairs. 1 kilobase pair is equal to 1000 base pairs.

"Mb": megabase pairs. 1 megabase pair is equal to 1000000 base pairs.

commas
A logical value indicating whether to include commas in start and stop labels. Default value is commas = TRUE.

sequence
A logical value indicating whether to include sequence information above the label of an x-axis (only at appropriate resolutions).

boxWidth
A numeric value indicating the width of the boxes representing sequence information at appropriate resolutions. Default value is boxWidth = 0.5.

axis
A character value indicating along which axis to add genome label. Sequence information will not be displayed along a y-axis. Default value is axis = "x". Options are:

"x": Genome label will be plotted along the x-axis.

"y": Genome label will be plotted along the y-axis.

at
A numeric vector of x-value locations for tick marks.

tcl
A numeric specifying the length of tickmarks as a fraction of text height. Default value is tcl = 0.5.

x
A numeric or unit object specifying genome label x-location.

y
A numeric or unit object specifying genome label y-location.

length
A numeric or unit object specifying length of genome label axis.

just
Justification of genome label relative to its (x, y) location. If there are two values, the first value specifies horizontal justification and the second value specifies vertical justification. Possible string values are: "left", "right", "centre", "center", "bottom", and "top". Default value is just = c("left", "top").

default.units
A string indicating the default units to use if x, y, or length are only given as numerics. Default value is default.units = "inches".

params
An optional bb_assembly object containing relevant function parameters.

...
Additional grid graphical parameters. See gpar.

Value

Returns a bb_genomeLabel object containing relevant genomic region, placement, and grob information.

Examples

## Load hg19 genomic annotation packages
library("TxDb.Hsapiens.UCSC.hg19.knownGene")
library("org.Hs.eg.db")

## Create BentoBox page
bb_pageCreate(width = 5, height = 4, default.units = "inches")

## Plot and place gene track on a BentoBox page
genesPlot <- bb_plotGenes(chrom = "chr8", chromstart = 1000000, chromend = 2000000, assembly = "hg19",
                          x = 0.5, y = 0.5, width = 4, height = 1.5, just = c("left", "top"), default.units = "inches")

## Add x-axis genome label with plot input
bb_plotGenomeLabel(plot = genesPlot, scale = "Kb",
                   x = 0.5, y = 2, just = c("left", "top"), default.units = "inches")

## Add x-axis genome label with chrom/chromstart/chromend input
bb_plotGenomeLabel(chrom = "chr8", chromstart = 1000000, chromend = 2000000, assembly = "hg19",
                   x = 0.5, y = 2.5, length = 4, just = c("left", "top"), default.units = "inches")