Plot data elements in a pileup style for a single chromosome

Plot data elements in a pileup style for a single chromosome

bb_plotPileup(data, chrom)
bb_plotPileup(data, chrom, x, y, width, height,
              just = c("left", "top"),
              default.units = "inches")

Arguments

data
Data to be plotted; as a character value specifying a BED file path, a data frame in BED format, a character value specifying a .bam file path where a bam index file (.bam.bai) is in the same directory, or a GRanges object.

chrom
Chromosome of region to be plotted, as a string.

chromstart
Integer start position on chromosome to be plotted.

chromend
Integer end position on chromosome to be plotted.

assembly
Default genome assembly as a string or a bb_assembly object. Default value is assembly = "hg19".

fill
Character value(s) as a single value, vector, or palette specifying fill colors of pileup elements. Default value is fill = "black".

colorby
A "colorby" object specifying information for scaling colors in data.

linecolor
A character value specifying the color of the lines outlining pileup elements. Default value is linecolor = NA.

boxHeight
A numeric or unit object specifying height of pileup element boxes. Default value is boxHeight = unit(2, "mm").

spaceWidth
A numeric value specifying the width of minimum spacing between pileup element boxes, as a fraction of the plot's genomic range. Default value is spaceWidth = 0.02.

spaceHeight
A numeric value specifying the height of spacing between pileup element boxes on different rows, as a fraction of boxHeight. Default value is spaceHeight = 0.3.

strandSplit
A logical value indicating whether plus and minus-stranded elements should be separated. Elements can only be split by strand if a strand column is found in data. Default value is strandSplit = FALSE.

bg
Character value indicating background color. Default value is bg = NA.

baseline
Logical value indicating whether to include a baseline along the x-axis. Default value is baseline = FALSE.

x
A numeric or unit object specifying pileup plot x-location.

y
A numeric or unit object specifying pileup plot y-location.

width
A numeric or unit object specifying pileup plot width.

height
A numeric or unit object specifying pileup plot height.

just
Justification of pileup plot relative to its (x, y) location. If there are two values, the first value specifies horizontal justification and the second value specifies vertical justification. Possible string values are: "left", "right", "centre", "center", "bottom", and "top". Default value is just = c("left", "top").

default.units
A string indicating the default units to use if x, y, width, or height are only given as numerics. Default value is default.units = "inches".

draw
A logical value indicating whether graphics output should be produced. Default value draw = TRUE.

params
An optional bb_assembly object containing relevant function parameters.

...
Additional grid graphical parameters. See gpar.

Value

Returns a bb_pileup object containing relevant genomic region, coloring data, placement, and grob information.

Examples

## Load BED data
data("bb_bedData")

## Plot pileup plot filling up entire graphic device
bb_plotPileup(data = bb_bedData, chrom = "chr21", chromstart = 29073000, chromend = 29074000, fill = "grey")

## Plot and place pileup plot on a BentoBox page
bb_pageCreate(width = 5, height = 5, default.units = "inches", xgrid = 0, ygrid = 0)
bb_plotPileup(data = bb_bedData, chrom = "chr21", chromstart = 29073000, chromend = 29074000, fill = "grey", baseline = TRUE,
             x = 0.5, y = 0.5, width = 4, height = 4, just = c("left", "top"), default.units = "inches")