Plot a polygon within a BentoBox layout

Plot a polygon within a BentoBox layout

bb_plotPolygon(x, y, default.units = "inches")

Arguments

x
A numeric vector or unit object specifying polygon vertex x-locations.

y
A numeric vector or unit object specifying polygon vertex y-locations.

default.units
A string indicating the default units to use if x or y are only given as numeric vectors. Default value is default.units = "inches".

linecolor
A character value specifying polygon line color. Default value is linecolor = "black".

lwd
A numeric specifying polygon line width. Default value is lwd = 1.

lty
A numeric specifying polygon line type. Default value is lty = 1.

fill
A character value specifying polygon fill color. Default value is fill = NA.

alpha
Numeric value specifying color transparency. Default value is alpha = 1.

id
A numeric vector used to separate locations in x and y into multiple polygons. All locations with the same id belong to the same polygon.

id.lengths
A numeric vector used to separate locations in x and y into multiple polygons. Specifies consecutive blocks of locations which make up separate polygons.

params
An optional bb_assembly object containing relevant function parameters.

...
Additional grid graphical parameters. See gpar.

Value

Returns a bb_polygon object containing relevant placement and grob information.

See also

Examples

## Create a BentoBox page
bb_pageCreate(width = 2, height = 2, default.units = "inches", xgrid = 0, ygrid = 0)

## Plot a 5-sided polygon
bb_plotPolygon(x = c(0.5, 1, 1.5, 1.25, 0.75), y = c(0.5, 0.25, 0.5, 1.25, 1.25), default.units = "inches")