Plot a rectangle within a BentoBox layout

Plot a rectangle within a BentoBox layout

bb_plotRect(x, y, width, height,
            just = "center",
            default.units = "inches")

Arguments

x
A numeric vector or unit object specifying rectangle x-locations.

y
A numeric vector or unit object specifying rectangle y-locations.

width
A numeric vector or unit object specifying rectangle widths.

height
A numeric vector or unit object specifying rectangle heights.

just
Justification of rectangle relative to its (x, y) location. If there are two values, the first value specifies horizontal justification and the second value specifies vertical justification. Possible string values are: "left", "right", "centre", "center", "bottom", and "top". Default value is just = "center".

default.units
A string indicating the default units to use if x, y, width, and height are only given as numerics or numeric vectors. Default value is default.units = "inches".

linecolor
A character value specifying rectangle line color. Default value is linecolor = "black".

lwd
A numeric specifying rectangle line width. Default value is lwd = 1.

lty
A numeric specifying rectangle line type. Default value is lty = 1.

fill
A character value specifying rectangle fill color. Default value is fill = NA.

alpha
Numeric value specifying color transparency. Default value is alpha = 1.

params
An optional bb_assembly object containing relevant function parameters.

...
Additional grid graphical parameters. See gpar.

Value

Returns a bb_rect object containing relevant placement and grob information.

See also

Examples

## Create a BentoBox page
bb_pageCreate(width = 2, height = 2, default.units = "inches", xgrid = 0, ygrid = 0)

## Plot a rectangle
bb_plotRect(x = 1, y = 1, width = 0.75, height = 0.25, just = "center", default.units = "inches")