Read a .hic file and return Hi-C data as a dataframe

Read a .hic file and return Hi-C data as a dataframe

bb_readHic(hicFile, chrom)

Arguments

hicFile
A character value specifying the path to the .hic file.

chrom
Chromosome of data, as a string.

chromstart
Integer start position on chromosome.

chromend
Integer end position on chromosome.

altchrom
Alternate chromosome for interchromosomal data, as a string.

altchromstart
Alternate chromosome integer start position for interchromosomal data.

altchromend
Alternate chromosome integer end position for interchromosomal data.

assembly
Default genome assembly as a string or a bb_assembly object. Default value is assembly = "hg19".

resolution
A numeric specifying the width of each pixel. "auto" will attempt to choose a resolution in basepairs based on the size of the region.

res_scale
A character value specifying the resolution scale. Default value is res_scale = "BP". Options are:

"BP": Base pairs.

"FRAG": Fragments.

zrange
A numeric vector of length 2 specifying the range of interaction scores, where extreme values will be set to the max or min.

norm
Character value specifying hic data normalization method. This value must be found in the .hic file. Default value is norm = "KR".

matrix
Character value indicating the type of matrix to output. Default value is matrix = "observed". Options are:

"observed": Observed counts.

"oe": Observed/expected counts.

params
An optional bb_assembly object containing relevant function parameters.

Value

Returns a 3-column dataframe in sparse upper triangular format with the following columns: chrom, altchrom, counts.

See also